requirements.txt
pytest==8.3.5
selenium==4.27.1
pytest-selenium==4.1.0
python-dotenv==1.0.1